Remove usage of UnitTestCase
