+++ tests/PHPUnit/Framework/TestCase/IntegrationTestCase.php
@@ -13,6 +13,7 @@ use Piwik\Config;
 use Piwik\Db;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Cache as PiwikCache;
+use Piwik\Tests\Framework\TestingEnvironmentVariables;
 
 /**
  * Tests extending IntegrationTestCase are much slower to run: the setUp will
@@ -77,7 +78,7 @@ abstract class IntegrationTestCase extends SystemTestCase
         self::$fixture->extraDefinitions = array_merge(static::provideContainerConfigBeforeClass(), $this->provideContainerConfig());
         self::$fixture->createEnvironmentInstance();
 
-        Fixture::loadAllPlugins(new \Piwik\Tests\Framework\TestingEnvironmentVariables(), get_class($this), self::$fixture->extraPluginsToLoad);
+        Fixture::loadAllPlugins(new TestingEnvironmentVariables(), get_class($this), self::$fixture->extraPluginsToLoad);
 
         Access::getInstance()->setSuperUserAccess(true);
