+++ tests/PHPUnit/Integration/TrackerTest.php
@@ -10,33 +10,15 @@ namespace Piwik\Tests\Integration;
 
 use Piwik\Common;
 use Piwik\Config;
-use Piwik\EventDispatcher;
 use Piwik\Piwik;
 use Piwik\Plugin;
 use Piwik\SettingsServer;
 use Piwik\Tests\Framework\Fixture;
 use Piwik\Tests\Framework\TestCase\IntegrationTestCase;
+use Piwik\Tests\Framework\Mock\Tracker\Handler;
+use Piwik\Tests\Framework\Mock\Tracker\RequestSet;
 use Piwik\Tracker;
-use Piwik\Tracker\RequestSet;
 use Piwik\Tracker\Request;
-
-class TestTracker extends Tracker
-{
-    public function __construct()
-    {
-        $this->isInstalled = true;
-    }
-
-    public function setIsNotInstalled()
-    {
-        $this->isInstalled = false;
-    }
-
-    public function disconnectDatabase()
-    {
-        parent::disconnectDatabase();
-    }
-}
 
 /**
  * @group TrackerTest
@@ -330,6 +312,53 @@ class TrackerTest extends IntegrationTestCase
                             $identifiedRequests[0]->getParams());
     }
 
+    public function test_main_shouldPostEndEvent()
+    {
+        $called = false;
+        Piwik::addAction('Tracker.end', function () use (&$called) {
+            $called = true;
+        });
+
+        $this->tracker->main(new Handler(), new RequestSet());
+
+        $this->assertTrue($called);
+    }
+
+    public function test_main_shouldPostEndEvent_EvenIfShouldNotRecordStats()
+    {
+        $called = false;
+        Piwik::addAction('Tracker.end', function () use (&$called) {
+            $called = true;
+        });
+
+        $handler = new Handler();
+
+        $this->tracker->disableRecordStatistics();
+        $this->tracker->main($handler, new RequestSet());
+
+        $this->assertFalse($handler->isProcessed);
+        $this->assertTrue($called);
+    }
+
+    public function test_main_shouldPostEndEvent_EvenIfThereIsAnException()
+    {
+        $called = false;
+        Piwik::addAction('Tracker.end', function () use (&$called) {
+            $called = true;
+        });
+
+        $handler = new Handler();
+        $handler->enableTriggerExceptionInProcess();
+
+        $requestSet = new RequestSet();
+        $requestSet->setRequests(array($this->buildRequest(1), $this->buildRequest(1)));
+
+        $this->tracker->main($handler, $requestSet);
+
+        $this->assertTrue($handler->isOnException);
+        $this->assertTrue($called);
+    }
+
     private function getDefaultHandler()
     {
         return new Tracker\Handler();
@@ -388,5 +417,36 @@ class TrackerTest extends IntegrationTestCase
         shell_exec("mv $localConfig $backupConfig");
         return array($localConfig, $backupConfig);
     }
+}
 
+class TestTracker extends Tracker
+{
+    public function __construct()
+    {
+        $this->isInstalled = true;
+        $this->record = true;
+    }
+
+    public function setIsNotInstalled()
+    {
+        $this->isInstalled = false;
+    }
+
+    public function disconnectDatabase()
+    {
+        parent::disconnectDatabase();
+    }
+
+    public function shouldRecordStatistics()
+    {
+        if (! $this->record) {
+            return false;
+        }
+        return parent::shouldRecordStatistics();
+    }
+
+    public function disableRecordStatistics()
+    {
+        $this->record = false;
+    }
 }
