+++ tests/PHPUnit/Unit/DataTable/Filter/ColumnCallbackDeleteMetadataTest.php
@@ -10,16 +10,13 @@ namespace Piwik\Tests\Core\DataTable\Filter;
 
 use Piwik\DataTable;
 use Piwik\DataTable\Row;
-use Piwik\Tests\Framework\TestCase\UnitTestCase;
 
 /**
  * @group ColumnCallbackDeleteMetadataTest
  * @group DataTable
  * @group Filter
- * @group Unit
- * @group Core
  */
-class ColumnCallbackDeleteMetadataTest extends UnitTestCase
+class ColumnCallbackDeleteMetadataTest extends \PHPUnit_Framework_TestCase
 {
     private $filter = 'ColumnCallbackDeleteMetadata';
