+++ tests/PHPUnit/Unit/DataTable/Filter/PrependSegmentFilterTest.php
@@ -10,16 +10,13 @@ namespace Piwik\Tests\Core\DataTable\Filter;
 
 use Piwik\DataTable;
 use Piwik\DataTable\Row;
-use Piwik\Tests\Framework\TestCase\UnitTestCase;
 
 /**
  * @group PrependSegmentTest
  * @group DataTable
  * @group Filter
- * @group Unit
- * @group Core
  */
-class PrependSegmentTest extends UnitTestCase
+class PrependSegmentTest extends \PHPUnit_Framework_TestCase
 {
     private $filter = 'PrependSegment';
