+++ tests/PHPUnit/Unit/DataTable/Filter/PrependValueToMetadataTest.php
@@ -10,16 +10,13 @@ namespace Piwik\Tests\Core\DataTable\Filter;
 
 use Piwik\DataTable;
 use Piwik\DataTable\Row;
-use Piwik\Tests\Framework\TestCase\UnitTestCase;
 
 /**
  * @group PrependValueToMetadataTest
  * @group DataTable
  * @group Filter
- * @group Unit
- * @group Core
  */
-class PrependValueToMetadataTest extends UnitTestCase
+class PrependValueToMetadataTest extends \PHPUnit_Framework_TestCase
 {
     private $filter = 'PrependValueToMetadata';
